The name of a controller, model or collection doesn't support names with / at start
The name of the controller can also be a relative path. But to make it uniform (just like with images) you should be able to add a `/` in front of the name/path. It currently doesn't support that. This fix resolves that for controllers,models and collections, and adds to the documentation it can also be a relative path.