refactor: remove chmodr
The file permissions that get set are ultimately only preserved when alloy creates the files in
Resources, when the SDK copies the files over to the application binary these ultimately become the
system defaults. Furthermore, files were incorrectly being given 755 permissions on first creation
and then changed to 644 on an incremental build.

We will now just use the node defaults for these files rather than setting them, files will be
given 644 and folders will be given 777